+++ cmd/frontend/internal/auth/userpasswd/reset_password.go
@@ -1,9 +1,11 @@
 package userpasswd
 
 import (
+	"context"
 	"encoding/json"
 	"net/http"
 
+	"github.com/pkg/errors"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/backend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/globals"
 	"github.com/sourcegraph/sourcegraph/internal/actor"
@@ -74,6 +76,60 @@ func HandleResetPasswordInit(w http.ResponseWriter, r *http.Request) {
 		httpLogAndError(w, "Could not send reset password email", http.StatusInternalServerError, "err", err)
 		return
 	}
+}
+
+var setPasswordEmailTemplates = txemail.MustValidate(txtypes.Templates{
+	Subject: `Set your Sourcegraph password`,
+	Text: `
+Your administrator created an account for you on Sourcegraph.
+
+To set the password for {{.Username}} on Sourcegraph, follow this link:
+
+  {{.URL}}
+`,
+	HTML: `
+<p>
+  Your administrator created an account for you on Sourcegraph.
+</p>
+
+<p><strong><a href="{{.URL}}">Set password for {{.Username}}</a></strong></p>
+`,
+})
+
+// HandleSetPasswordEmail sends the password reset email directly to the user for users created by site admins.
+func HandleSetPasswordEmail(ctx context.Context, id int32) (string, error) {
+	e, _, err := db.UserEmails.GetPrimaryEmail(ctx, id)
+	if err != nil {
+		return "", errors.Wrap(err, "get user primary email")
+	}
+
+	usr, err := db.Users.GetByID(ctx, id)
+	if err != nil {
+		return "", errors.Wrap(err, "get user by ID")
+	}
+
+	ru, err := backend.MakePasswordResetURL(ctx, id)
+	if err == db.ErrPasswordResetRateLimit {
+		return "", err
+	} else if err != nil {
+		return "", errors.Wrap(err, "make password reset URL")
+	}
+
+	rus := globals.ExternalURL().ResolveReference(ru).String()
+	if err := txemail.Send(ctx, txemail.Message{
+		To:       []string{e},
+		Template: setPasswordEmailTemplates,
+		Data: struct {
+			Username string
+			URL      string
+		}{
+			Username: usr.Username,
+			URL:      rus,
+		},
+	}); err != nil {
+		return "", err
+	}
+	return rus, nil
 }
 
 var resetPasswordEmailTemplates = txemail.MustValidate(txtypes.Templates{